Working with the User Authentication Status

+++ app/Http/Controllers/PostController.php
@@ -23,6 +23,10 @@ class PostController extends Controller
 
     public function getAdminIndex()
     {
+        if(!Auth::check())
+        {
+            return redirect()->back();
+        }
         $posts = Post::orderBy("title", 'asc')->get();
         return view('admin.index', ['posts' => $posts]);
     }
@@ -45,12 +49,20 @@ class PostController extends Controller
 
     public function getAdminCreate()
     {
+        if(!Auth::check())
+        {
+            return redirect()->back();
+        }
         $tags = Tag::all();
         return view('admin.create', ['tags' => $tags]);
     }
 
     public function getAdminEdit( $id)
     {
+        if(!Auth::check())
+        {
+            return redirect()->back();
+        }
        $post = Post::find($id);
        $tags = Tag::all();
        return view('admin.edit', ['post' => $post, 'postId' => $id, 'tags' => $tags]);
@@ -88,6 +100,10 @@ class PostController extends Controller
 
     public function postAdminUpdate(Request $request)
     {
+        if(!Auth::check())
+        {
+            return redirect()->back();
+        }
         $this->validate($request, [
             'title' => 'required|min:5',
             'content' => 'required|min:10'
@@ -107,6 +123,10 @@ class PostController extends Controller
 
     public function getAdminDelete($id)
     {
+        if(!Auth::check())
+        {
+            return redirect()->back();
+        }
         $post = Post::find($id);
         $post->likes()->delete();
         $post->tags()->detach();
